+++ dist-staging/assets/iframe.5b8b978e.js
@@ -1,1 +1,0 @@
-import{d as r,a as n,e as o,o as a,g as s}from"./vendor.98fdfd48.js";const i=["height"],l=r({setup(c){const e=n(),t=n(0);return o(()=>{t.value=e.value.parentElement.clientHeight}),(m,f)=>(a(),s("iframe",{ref_key:"myFrame",ref:e,src:"http://qingqingxuan.gitee.io/work-p-site",width:"100%",height:t.value+"px",frameborder:"0"},null,8,i))}});export{l as default};